Split env effect rejection fixtures
