--- tests/integration/cli_build/declared_env_effects/rejections.rs
@@ -1,423 +1,0 @@
-use std::process::Command;
-
-use super::*;
-
-#[test]
-fn build_command_build_zen_rejects_env_read_without_fallback_before_execution() {
-    assert_env_read_without_fallback_is_rejected(
-        &["build", "build.zen"],
-        "build command should not start after graph validation fails",
-    );
-}
-
-#[test]
-fn build_command_multi_target_build_zen_rejects_env_read_without_fallback_before_execution() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) { value }
-    b.add(Executable { name: "app", main: "app.zen", out_dir: "build/app/" })
-    b.add(Executable { name: "tool", main: "tool.zen", out_dir: "build/tool/" })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["build", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen build build.zen");
-
-    assert_env_read_without_fallback_failed(&output);
-    assert!(
-        !tmp.path().join("build").exists(),
-        "build command should reject env effects before target execution"
-    );
-}
-
-#[test]
-fn build_command_build_zen_rejects_env_read_without_fallback_before_unselected_targets() {
-    assert_env_read_without_fallback_before_unselected_targets(
-        &["build", "build.zen"],
-        "build command should reject env effects before selected target execution",
-    );
-}
-
-fn assert_env_read_without_fallback_before_unselected_targets(args: &[&str], build_message: &str) {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) { value }
-    b.add(Executable { name: "app", main: "app.zen", out_dir: "build/app/" })
-    b.add(Test { name: "unit", root: "missing_unit.zen" })
-    b.add(Library { name: "core", exports: ["lib.zen"] })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-    std::fs::write(
-        tmp.path().join("app.zen"),
-        r#"
-main = () i32 {
-    0
-}
-"#,
-    )
-    .expect("write app.zen");
-    std::fs::write(
-        tmp.path().join("lib.zen"),
-        r#"
-value = () i32 {
-    1
-}
-"#,
-    )
-    .expect("write lib.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(args)
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen executable build graph command");
-
-    assert_env_read_without_fallback_failed(&output);
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        !stderr.contains("missing_unit.zen"),
-        "host-effect validation should run before unrelated test source handling, stderr={stderr}"
-    );
-    assert!(!tmp.path().join("build").exists(), "{build_message}");
-}
-
-#[test]
-fn direct_file_command_build_zen_rejects_env_read_without_fallback_before_execution() {
-    assert_env_read_without_fallback_is_rejected(
-        &["build.zen"],
-        "direct build.zen command should not start after graph validation fails",
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_rejects_env_read_without_fallback_before_unselected_targets() {
-    assert_env_read_without_fallback_before_unselected_targets(
-        &["build.zen"],
-        "direct build.zen command should reject env effects before selected target execution",
-    );
-}
-
-#[test]
-fn direct_file_command_multi_target_build_zen_rejects_env_read_without_fallback_before_execution() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) { value }
-    b.add(Executable { name: "app", main: "app.zen", out_dir: "build/app/" })
-    b.add(Executable { name: "tool", main: "tool.zen", out_dir: "build/tool/" })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .arg("build.zen")
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen build.zen");
-
-    assert_env_read_without_fallback_failed(&output);
-    assert!(
-        !tmp.path().join("build").exists(),
-        "direct build.zen command should reject env effects before multi-target execution"
-    );
-}
-
-#[test]
-fn build_graph_command_rejects_env_read_without_fallback_before_execution() {
-    assert_env_read_without_fallback_is_rejected(
-        &["build-graph", "build.zen"],
-        "build-graph command should not start after graph validation fails",
-    );
-}
-
-#[test]
-fn build_graph_command_multi_target_rejects_env_read_without_fallback_before_execution() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) { value }
-    b.add(Executable { name: "app", main: "app.zen", out_dir: "build/app/" })
-    b.add(Executable { name: "tool", main: "tool.zen", out_dir: "build/tool/" })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["build-graph", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen build-graph build.zen");
-
-    assert_env_read_without_fallback_failed(&output);
-    assert!(
-        !tmp.path().join("build").exists(),
-        "build-graph command should reject env effects before target execution"
-    );
-}
-
-#[test]
-fn check_command_build_zen_rejects_env_read_without_fallback_before_source_validation() {
-    let tmp = executable_graph_with_env_read_without_fallback("missing.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["check", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen check build.zen");
-
-    assert_env_read_without_fallback_failed(&output);
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        !stderr.contains("source not found"),
-        "host-effect validation should run before source validation, stderr={stderr}"
-    );
-}
-
-#[test]
-fn check_command_multi_target_build_zen_rejects_env_read_without_fallback() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) { value }
-    b.add(Executable { name: "app", main: "missing_app.zen", out_dir: "build/app/" })
-    b.add(Test { name: "unit", root: "missing_unit.zen" })
-    b.add(Library { name: "core", exports: ["missing_lib.zen"] })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["check", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen check build.zen");
-
-    assert_env_read_without_fallback_failed(&output);
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        !stderr.contains("source not found"),
-        "host-effect validation should run before multi-target source validation, stderr={stderr}"
-    );
-}
-
-#[test]
-fn emit_command_build_zen_rejects_env_read_without_fallback() {
-    let tmp = executable_graph_with_env_read_without_fallback("main.zen");
-    std::fs::write(
-        tmp.path().join("main.zen"),
-        r#"
-main = () i32 {
-    0
-}
-"#,
-    )
-    .expect("write main.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["emit", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen emit build.zen");
-
-    assert_env_read_without_fallback_failed(&output);
-    assert!(
-        output.stdout.is_empty(),
-        "emit should not write C source after graph validation fails, stdout={}",
-        String::from_utf8_lossy(&output.stdout)
-    );
-}
-
-#[test]
-fn emit_command_build_zen_rejects_env_read_without_fallback_before_unselected_targets() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) { value }
-    b.add(Executable { name: "app", main: "app.zen", out_dir: "build/app/" })
-    b.add(Test { name: "unit", root: "missing_test.zen" })
-    b.add(Library { name: "core", exports: ["missing_lib.zen"] })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-    std::fs::write(
-        tmp.path().join("app.zen"),
-        r#"
-main = () i32 {
-    0
-}
-"#,
-    )
-    .expect("write app.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["emit", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen emit build.zen");
-
-    assert_env_read_without_fallback_failed(&output);
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        !stderr.contains("source not found"),
-        "emit should reject env effects before unselected target source validation, stderr={stderr}"
-    );
-    assert!(
-        output.stdout.is_empty(),
-        "emit should not write C source after graph validation fails, stdout={}",
-        String::from_utf8_lossy(&output.stdout)
-    );
-    assert!(
-        !tmp.path().join("build").exists(),
-        "zen emit build.zen should not create build outputs"
-    );
-}
-
-#[test]
-fn test_command_build_zen_rejects_env_read_without_fallback_before_execution() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) { value }
-    b.add(Test { name: "unit", root: "test.zen" })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["test", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen test build.zen");
-
-    assert_env_read_without_fallback_failed(&output);
-    assert!(
-        !tmp.path().join("build").exists(),
-        "test command should not start after graph validation fails"
-    );
-}
-
-#[test]
-fn test_command_build_zen_rejects_env_read_without_fallback_before_unselected_targets() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) { value }
-    b.add(Executable { name: "app", main: "missing_app.zen", out_dir: "build/app/" })
-    b.add(Test { name: "unit", root: "unit.zen" })
-    b.add(Library { name: "core", exports: ["lib.zen"] })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-    std::fs::write(
-        tmp.path().join("unit.zen"),
-        r#"
-main = () i32 {
-    0
-}
-"#,
-    )
-    .expect("write unit.zen");
-    std::fs::write(
-        tmp.path().join("lib.zen"),
-        r#"
-value = () i32 {
-    1
-}
-"#,
-    )
-    .expect("write lib.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["test", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen test build.zen");
-
-    assert_env_read_without_fallback_failed(&output);
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        !stderr.contains("missing_app.zen"),
-        "host-effect validation should run before unrelated executable source handling, stderr={stderr}"
-    );
-    assert!(
-        !tmp.path().join("build").exists(),
-        "test command should not start after graph validation fails"
-    );
-}
-
-#[test]
-fn test_command_multi_target_build_zen_rejects_env_read_without_fallback_before_execution() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) { value }
-    b.add(Test { name: "unit", root: "unit.zen" })
-    b.add(Test { name: "integration", root: "integration.zen" })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["test", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen test build.zen");
-
-    assert_env_read_without_fallback_failed(&output);
-    assert!(
-        !tmp.path().join("build").exists(),
-        "test command should reject env effects before multi-target execution"
-    );
-}
